requirements.txt
# ============================================================================
# Python Dependencies
# ============================================================================
# Install: pip install -r requirements.txt
# ============================================================================

# Data generation
faker>=22.0.0
pandas>=2.0.0
numpy>=1.24.0

# Database
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0

# dbt
dbt-postgres>=1.7.0